fixed adding costs and payments

See merge request Toni222/Eureka!91

+++ backend/src/main/java/hr/fer/opp/eureka/service/BuildingService.java
@@ -23,5 +23,5 @@ public interface BuildingService {
 
   Building editFunds(Building building);
 
-  Building editFundsForUser(Long amount, Long userId);
+  Building editFundsForUser(Float amount, Long userId);
 }
